Add files via upload
Changed all floats to doubles to increase precision. 

Added time array which carried through to minorframes (time could just be inferred from the number of bytes in from there)

+++ POESTIPdemod/AGC.c
@@ -1,13 +1,13 @@
 #include <complex.h>
 #include <math.h>
 #include "AGC.h"
-float StaticGain(double complex *complexData,unsigned int nSamples,float desiredLevel)
+double StaticGain(double complex *complexData,unsigned int nSamples,double desiredLevel)
    {
    ////n=100000; //number of points to average
    //if nargin < 2 || n==0 
        //n = numel(dataStreamIn);
    int i;
-   float avgLevel=0;
+   double avgLevel=0;
    avgLevel = cabs(complexData[0]);
    for(i=1; i < nSamples; i++)
       {
@@ -20,16 +20,16 @@ float StaticGain(double complex *complexData,unsigned int nSamples,float desired
    }
    
 //Automatic Gain Control Block
-void NormalizingAGC(float *dataStreamIn, unsigned long nSamples, double AGC_loop_gain)
+void NormalizingAGC(double *dataStreamIn, unsigned long nSamples, double AGC_loop_gain)
    {
    //Todo: implement a 'relock' mode for LARGE error values (either adjust gain
    //outright or adjust loop gain
    
    unsigned long idx;
    
-   static float gain = 1; //Initial Gain Value
-   float desired = 0.6366; //because a sin wave of amplitude 1 has this average absolute value
-   float error;
+   static double gain = 1; //Initial Gain Value
+   double desired = 0.6366; //because a sin wave of amplitude 1 has this average absolute value
+   double error;
    
    for(idx=0; idx< nSamples; idx++)
       {
