+++ POESTIPdemod/AGC.h
@@ -1,6 +1,6 @@
 #ifndef AGC_H
 #define AGC_H
 
-float StaticGain(double complex *complexData,unsigned int nSamples,float desiredLevel);
-void NormalizingAGC(float *dataStreamIn, unsigned long nSamples, double AGC_loop_gain);
+double StaticGain(double complex *complexData,unsigned int nSamples,double desiredLevel);
+void NormalizingAGC(double *dataStreamIn, unsigned long nSamples, double AGC_loop_gain);
 #endif
